+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **@oneblink/uploads (NPM package)** to product repositories
+- **@oneblink/storage (NPM package)** to product repositories
 
 ## [3.1.0] - 2024-03-26