+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.2] - 2024-08-07
+
+##### Release Name: Navy Spheres
+
 ### Added
 
 - `product-s3-object-lambda` to repo list